DO NOT MERGE vision: logo detection GA samples
