+++ video/cloud-client/analyze/requirements.txt
@@ -1,2 +1,2 @@
-google-cloud-videointelligence==1.13.0
+google-cloud-videointelligence==1.14.0
 google-cloud-storage==1.26.0
